requirements.txt
# AI and ML
langchain>=0.1.0
openai>=1.0.0
anthropic>=0.10.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Vector Databases
weaviate-client>=3.20.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Email
email-validator>=2.0.0
imaplib2>=3.0.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
